+++ docs/releases.md
@@ -5,8 +5,10 @@
 
 * [FEATURE] The selector `ZeroClipboard.$` was removed in favor of leaving it up to the user. #63
 * [FEATURE] Creating new instance of the class is easier `new ZeroClipboard()` #65
+* [FEATURE] Overhaul of the options setting and getting. Now via the constructor or setDefaults.
 * [BUG] It was possible to glue an element multiple times.
 * [BUG] Removing multiple functions for same action. Was causing multiple firings.
+* [BUG] After destroy, there was an error on mouseover of old client.
 
 ### ZeroClipboard 1.1.5